requirements.txt
Flask>=3.0
gunicorn>=23.0
PyYAML>=6.0
